+++ BitForce_SC/ASIC_Engine.h
@@ -117,7 +117,7 @@ int     ASIC_get_job_status(unsigned int *iNonceList, char *iNonceCount);
 void    ASIC_job_issue(void* pJobPacket, unsigned int _LowRange,unsigned int _HighRange);
 void	ASIC_job_issue_p2p(void* pJobPacketP2P);
 int  	ASIC_get_chip_count(void);
-int		ASIC_are_all_engines_done(unsigned short iChip);
+int		ASIC_are_all_engines_done(unsigned int iChip);
 int		ASIC_does_chip_exist(unsigned int iChipIndex);
 int		ASIC_is_processing(void);
 void	ASIC_Bootup_Chips(void);
